migrate id's from int to long (#104)

+++ src/Hangfire.PostgreSql/Entities/JobParameter.cs
@@ -26,7 +26,7 @@ namespace Hangfire.PostgreSql.Entities
     [UsedImplicitly]
     internal class JobParameter
     {
-        public int JobId { get; set; }
+        public long JobId { get; set; }
         public string Name { get; set; }
         public string Value { get; set; }
     }
